+++ components/experimental/vltava/src/interfaces/componentInternalRegistry.ts
@@ -1,0 +1,37 @@
+/*!
+ * Copyright (c) Microsoft Corporation. All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+import { IComponent } from "@fluidframework/component-core-interfaces";
+import { IProvideComponentFactory } from "@fluidframework/runtime-definitions";
+
+declare module "@fluidframework/component-core-interfaces" {
+    // eslint-disable-next-line @typescript-eslint/no-empty-interface
+    export interface IComponent extends Readonly<Partial<IProvideComponentInternalRegistry>> { }
+}
+
+export const IComponentInternalRegistry: keyof IProvideComponentInternalRegistry = "IComponentInternalRegistry";
+
+export interface IProvideComponentInternalRegistry {
+    readonly IComponentInternalRegistry: IComponentInternalRegistry;
+}
+
+/**
+ * Provides functionality to retrieve subsets of an internal registry.
+ */
+export interface IComponentInternalRegistry extends IProvideComponentInternalRegistry {
+    getFromCapability(type: keyof IComponent): IInternalRegistryEntry[];
+    hasCapability(type: string, capability: keyof IComponent): boolean;
+}
+
+/**
+ * A registry entry, with extra metadata.
+ */
+export interface IInternalRegistryEntry {
+    type: string;
+    factory: Promise<IProvideComponentFactory>;
+    capabilities: (keyof IComponent)[];
+    friendlyName: string;
+    fabricIconName: string;
+}
